telegram_bot/requirements.txt
python-telegram-bot==21.6
pymongo==4.6.3
python-dotenv==1.0.1
